test: refactor to take manual test out in its own separated file
