+++ packages/tracing/tests/e2e/constants.ts
@@ -1,8 +1,8 @@
 export const RESOURCE_NAME_PREFIX = 'Tracer-E2E';
-export const ONE_MINUTE = 60 * 10_00;
+export const ONE_MINUTE = 60 * 1_000;
 export const TEST_CASE_TIMEOUT = ONE_MINUTE * 2;
-export const SETUP_TIMEOUT = 300_000;
-export const TEARDOWN_TIMEOUT = 200_000;
+export const SETUP_TIMEOUT = ONE_MINUTE * 5;
+export const TEARDOWN_TIMEOUT = ONE_MINUTE * 3;
 
 export const expectedCustomAnnotationKey = 'myAnnotation';
 export const expectedCustomAnnotationValue = 'myValue';
@@ -10,7 +10,6 @@ export const expectedCustomMetadataKey = 'myMetadata';
 export const expectedCustomMetadataValue = { bar: 'baz' };
 export const expectedCustomResponseValue = { foo: 'bar' };
 export const expectedCustomErrorMessage = 'An error has occurred';
-
 
 /**
  * A type that contains information for invoking a Lambda function,
